Add files via upload
Hey everyone, I uploaded my version of the code. The only part missing is the conclusion. I also uploaded my handwritten work to check for problem 5 and the bonus. Based off the results, I would say the model is not accurate in either case as the predicted values and model values do not overlap. Happy to answer questions, hope this helps!